Pulled AgentShell into its own file.
